# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WEMU API Documentation",
        "description": "Swagger documentation for all APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/docs",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/recently-played": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get recently played songs",
                "operationId": "f65f9c900dc99d4ca85ce9f894799007",
                "responses": {
                    "200": {
                        "description": "Songs fetched successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/toggle-song-like/{songId}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Like or unlike a song",
                "operationId": "41086265806c05fbb1d1dd826f7bc7fb",
                "parameters": [
                    {
                        "name": "songId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/liked-songs": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get liked songs",
                "operationId": "ffd206a0c2b9cff164dcdf6f45164517",
                "responses": {
                    "200": {
                        "description": "Songs fetched successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/made-for-you": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get made for you songs",
                "operationId": "13aaf03efcd1cb1c9ab092436b5f5918",
                "responses": {
                    "200": {
                        "description": "Songs fetched successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/toggle-artist-follow/{artistId}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Follow or unfollow an artist",
                "operationId": "04ad6a8184109300d7aaab97a576efc4",
                "parameters": [
                    {
                        "name": "artistId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Follow status updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favourite-artists": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get favourite artists",
                "operationId": "cebd7d28841e43d761a828dd1a3a371b",
                "responses": {
                    "200": {
                        "description": "Artists fetched successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/toggle-artist-preference": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Toggle artist preference",
                "operationId": "ee6d2d1d6de7457830e81495573570aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "artist_id"
                                ],
                                "properties": {
                                    "artist_id": {
                                        "description": "Comma separated artist IDs",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preference updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}